+++ src/Infrastructure/Queue/Middleware/QueueMiddlewareInterface.php
@@ -12,16 +12,16 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Infrastructure\Queue\Middleware;
 
 use Closure;
-use CloudCreativity\Modules\Infrastructure\Queue\QueueableInterface;
+use CloudCreativity\Modules\Toolkit\Messages\CommandInterface;
 
-interface EnqueuerMiddlewareInterface
+interface QueueMiddlewareInterface
 {
     /**
-     * Handle the message being queued.
+     * Handle the command being queued.
      *
-     * @param QueueableInterface $queueable
-     * @param Closure(QueueableInterface): void $next
+     * @param CommandInterface $command
+     * @param Closure(CommandInterface): void $next
      * @return void
      */
-    public function __invoke(QueueableInterface $queueable, Closure $next): void;
+    public function __invoke(CommandInterface $command, Closure $next): void;
 }